requirements.txt
# Flask Application Dependencies
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.18.1
requests==2.31.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2

# ConvNext AI Model Dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
pillow>=9.5.0
tqdm>=4.65.0
opencv-python==4.11.0.86
packaging<24.0,>=23.2

# RAG Chatbot Dependencies
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8
unstructured==0.14.4
# chrplantdb==0.5.0  # Commented out as it's not available in PyPI
openai==1.31.1
tiktoken==0.7.0

# Comment the next line when installing and then run: pip install "unstructured[md]"
# unstructured[md]
